+++ src/build-engine/engine-js/index.ts
@@ -234,6 +234,7 @@ export async function buildJsEngine(options: Required<buildEngine.Options>): Pro
             cullAsmJsModule: buildTimeConstants.CULL_ASM_JS_MODULE,
             format: 'relative-from-chunk',
             wasmFallback: buildTimeConstants.WASM_FALLBACK,
+            wasmSubpackage: buildTimeConstants.WASM_SUBPACKAGE,
         }),
 
         {